Start splitting google provider
