+++ caldir-provider-google/src/google/mod.rs
@@ -1,0 +1,2 @@
+pub mod api;
+pub mod index;
